# Conflicts:
#	RR-Import/src/main/java/com/scor/rr/configuration/ImportLossDataJob.java
#	RR-Import/src/main/java/com/scor/rr/domain/riskLink/RlSourceResult.java
#	RR-Import/src/test/java/RiskRevealApplicationTests.java

+++ RR-Import/src/main/java/com/scor/rr/service/batch/ELTToPLTConverter.java
@@ -121,7 +121,7 @@ public class ELTToPLTConverter {
         log.debug("Modeling basis: {}", modelingBasis);
 
         Map<String, List<PLTHeader>> pltsByPeqt = new HashMap<>();
-        Map<Integer, Map> bundleForPLT = new HashMap<>();
+        Map<Long, Map> bundleForPLT = new HashMap<>();
 
         for (Map bundle : bundles) {
 
@@ -260,7 +260,7 @@ public class ELTToPLTConverter {
                 int end = (i + 1) * chunkSize < entry.getValue().size() ? (i + 1) * chunkSize : entry.getValue().size();
                 List<PLTHeader> scorPLTHeaders = entry.getValue().subList(start, end);
 //                List<ScorPLTHeader> scorPLTHeaders = entry.getValue();
-                Map<Integer, Map<Long, ELTLossBetaConvertFunction>> convertFunctionMapForPLT = new HashMap<>(scorPLTHeaders.size());
+                Map<Long, Map<Long, ELTLossBetaConvertFunction>> convertFunctionMapForPLT = new HashMap<>(scorPLTHeaders.size());
                 for (PLTHeader scorPLTHeader : scorPLTHeaders) {
                     Map bundle = bundleForPLT.get(scorPLTHeader.getPltHeaderId());
                     if (bundle != null) {
@@ -453,12 +453,12 @@ public class ELTToPLTConverter {
         private CountDownLatch latch;
         private BinFile peqtFile;
         private List<PLTHeader> scorPLTHeaders;
-        Map<Integer, Map<Long, ELTLossBetaConvertFunction>> convertFunctionMapForPLT;
+        Map<Long, Map<Long, ELTLossBetaConvertFunction>> convertFunctionMapForPLT;
         private Boolean[] finished = { Boolean.FALSE };
         private int nbWorkers;
 
         public TreatyBatchLauncher(CountDownLatch latch, BinFile peqtFile, List<PLTHeader> scorPLTHeaders,
-                                   Map<Integer, Map<Long, ELTLossBetaConvertFunction>> convertFunctionMapForPLT) {
+                                   Map<Long, Map<Long, ELTLossBetaConvertFunction>> convertFunctionMapForPLT) {
             this.id = hashCode();
             this.queue = new ArrayBlockingQueue<>(queueSize);
             this.latch = latch;
@@ -495,7 +495,7 @@ public class ELTToPLTConverter {
             CountDownLatch workerLatch = new CountDownLatch(nbWorkers);
             ExecutorService pool = Executors.newFixedThreadPool(nbWorkers);
             Timer timer = new Timer();
-            Map<Integer, FileOutputStream> outputStreamForPLT = new HashMap<>();
+            Map<Long, FileOutputStream> outputStreamForPLT = new HashMap<>();
             for (PLTHeader pltHeader : scorPLTHeaders) {
                 File file = new File(pltHeader.getPltLossDataFilePath(), pltHeader.getPltLossDataFileName());
                 try {
@@ -607,7 +607,7 @@ public class ELTToPLTConverter {
                 log.error("Exception: {}, {}", peqtFile.getFileName(), e);
             }
 
-            for (Map.Entry<Integer, FileOutputStream> entry : outputStreamForPLT.entrySet()) {
+            for (Map.Entry<Long, FileOutputStream> entry : outputStreamForPLT.entrySet()) {
                 IOUtils.closeQuietly(entry.getValue());
             }
 
@@ -642,15 +642,15 @@ public class ELTToPLTConverter {
         private final BlockingQueue<RRPeriod> queue;
         private final CountDownLatch workerLatch;
         private List<PLTHeader> scorPLTHeaders;
-        private Map<Integer, Map<Long, ELTLossBetaConvertFunction>> convertFunctionMapForPLT;
-        private Map<Integer, FileOutputStream> outputStreamForPLT;
+        private Map<Long, Map<Long, ELTLossBetaConvertFunction>> convertFunctionMapForPLT;
+        private Map<Long, FileOutputStream> outputStreamForPLT;
         private Timer timer;
         private Comparator<PLTLossData> cmp;
-        private Map<Integer, List<PLTLossData>> pltLossDataForPLT;
+        private Map<Long, List<PLTLossData>> pltLossDataForPLT;
 
 
         public TreatyBatchWorker(int masterId, Boolean[] finished, BlockingQueue<RRPeriod> queue, CountDownLatch workerLatch, Timer timer, List<PLTHeader> scorPLTHeaders,
-                                 Map<Integer, Map<Long, ELTLossBetaConvertFunction>> convertFunctionMapForPLT, Map<Integer, FileOutputStream> outputStreamForPLT) {
+                                 Map<Long, Map<Long, ELTLossBetaConvertFunction>> convertFunctionMapForPLT, Map<Long, FileOutputStream> outputStreamForPLT) {
             id = hashCode();
             this.masterId = masterId;
             this.finished = finished;
